+++ CHANGELOG.md
@@ -21,6 +21,7 @@ permalink: /docs/en-US/changelog/
 * Fixed an issue with `composer create-project` not running when specified in `config.yml` ( #2565 )
 * Switched obsolete mirror check for MariaDB to the one already used (#2575)
 * Fixed a broken warning in the network checks
+* Fixed an issue with `/root/.local/share/composer` when provisioning (#2589)
 
 ## 3.8.1 ( 2021 November 15th )